+++ semgrep-core/tests/ts/parsing/typeof.ts
@@ -1,0 +1,2 @@
+type LinksKeys = keyof typeof DOCS_LINKS;
+type LinksValues = typeof DOCS_LINKS[LinksKeys];